// src/features/EditableProfileCard/model/types/EditableProfileCardSchema.ts
import { Profile } from 'entities/Profile';
import { ValidateProfileError } from '../consts/consts';

export interface EditableProfileCardSchema {
    data?: Profile;
    form?: Profile;
    isLoading: boolean;
    error?: string;
    readonly: boolean;
    validateErrors?: ValidateProfileError[];
}